feature: Users can save or discard avatar changes
The avatar feature’s save button actually works now. Also, if
users make changes to the avatar but don’t save it, a pop up
will show, asking if they want to discard their changes. This is
not connected to any database. So avatar doesn’t save to a
user’s account.